move weighted_mean to quantity
