+++ propnet/core/materials.py
@@ -5,9 +5,8 @@ Module containing classes and methods for Material functionality in propnet code
 from collections import defaultdict
 from itertools import chain
 
-from propnet.core.quantity import Quantity
+from propnet.core.quantity import Quantity, weighted_mean
 from propnet.core.symbols import Symbol
-from propnet.core.utils import weighted_mean
 
 
 class Material(object):
