+++ propnet/core/quantity.py
@@ -1,11 +1,14 @@
 # TODO: remove typing
 from typing import *
 
+import numpy as np
 from monty.json import MSONable
 
 from propnet import ureg
 from propnet.core.symbols import Symbol
+from propnet.core.provenance import ProvenanceElement
 from propnet.symbols import DEFAULT_SYMBOLS
+from uncertainties import unumpy
 
 
 class Quantity(MSONable):
@@ -121,3 +124,58 @@ class Quantity(MSONable):
                 "units": units.format_babel() if units else None,
                 "@module": "propnet.core.quantity",
                 "@class": "Quantity"}
+
+
+def weighted_mean(quantities):
+    """
+    Function to retrieve weighted mean
+
+    Args:
+        quantities ([Quantity]): list of quantities
+
+    Returns:
+        weighted mean
+    """
+    # can't run this twice yet ...
+    # TODO: remove
+    if hasattr(quantities[0].value, "std_dev"):
+        return quantities
+
+    input_symbol = quantities[0].symbol
+    if input_symbol.category == 'object':
+        # TODO: can't average 'objects', highlights a weakness in Quantity class
+        # would be fixed by changing this class design ...
+        return quantities
+
+    if not all(input_symbol == q.symbol for q in quantities):
+        raise ValueError("Can only calculate a weighted mean if all quantities "
+                         "refer to the same symbol.")
+
+    # TODO: an actual weighted mean; just a simple mean at present
+    # TODO: support propagation of uncertainties (this will only work once at present)
+
+    # TODO: test this with units, not magnitudes ... remember units may not be canonical units(?)
+    if isinstance(quantities[0].value, list):
+        # hack to get arrays working for now
+        vals = [q.value for q in quantities]
+    else:
+        vals = [q.value.magnitude for q in quantities]
+
+    new_magnitude = np.mean(vals, axis=0)
+    std_dev = np.std(vals, axis=0)
+    new_value = unumpy.uarray(new_magnitude, std_dev)
+
+    new_tags = set()
+    new_provenance = ProvenanceElement(model='aggregation', inputs=[])
+    for q in quantities:
+        if q.tags:
+            for t in q.tags:
+                new_tags.add(t)
+        new_provenance.inputs.append(q)
+
+    new_quantity = Quantity(symbol_type=input_symbol,
+                            value=new_value,
+                            tags=list(new_tags),
+                            provenance=new_provenance)
+
+    return new_quantity
